[linear] Fix bugs
The first fixed bug concerns the translation of the Egoto instruction.
We forgot to mark the destination label as needed during translation.

The second fixed bug concerns the translation of the Emubranch
instructions. The x86-64 instruction "setcc" SETS ONLY A BYTE.
Therefore if we give it a 64-bit register, for example, only
the first byte (in a little-endian machine, at least) will be
modified. Therefore, if we want to use the result of a "setcc"
instruction we must use a 8-bit instruction (movb, testb, e.g.).